Split session view task-run helpers
