+++ apps/desktop/src/lib/session-view-tokens.ts
@@ -1,0 +1,19 @@
+import type { SessionTokens } from '@open-cowork/shared'
+
+export const EMPTY_SESSION_TOKENS: SessionTokens = {
+  input: 0,
+  output: 0,
+  reasoning: 0,
+  cacheRead: 0,
+  cacheWrite: 0,
+}
+
+export function cloneTokens(tokens: SessionTokens): SessionTokens {
+  return {
+    input: tokens.input,
+    output: tokens.output,
+    reasoning: tokens.reasoning,
+    cacheRead: tokens.cacheRead,
+    cacheWrite: tokens.cacheWrite,
+  }
+}
